+++ packages/frontend/src/types/related-twreporter-item.ts
@@ -1,0 +1,6 @@
+export type RelatedType = 'www-topic' | 'www-article'
+
+export type RelatedItem = {
+  type: RelatedType
+  slug: string
+}